updating hot_fix branch with main branch changes

+++ handlers/Instructor.go
@@ -6,14 +6,37 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
+	"github.com/google/uuid"
 )
 
+type res struct {
+	Msg string
+	Err string
+	URl string
+}
+
 func (h *Handler) InstructorInfoHandlers(ctx *gin.Context) {
-	type res struct {
-		Msg string
-		Err string
-		URl string
+
+	token, err3 := ctx.Cookie("token")
+	if err3 != nil {
+		ctx.JSON(http.StatusInternalServerError, err3.Error())
+		return
 	}
+	token_id, err4 := uuid.Parse(token)
+	if err4 != nil {
+		ctx.JSON(http.StatusInternalServerError, fmt.Errorf("error parsing uuid").Error())
+		return
+	}
+	status, err1 := h.service.CheckTokenValidity(token_id)
+	if err1 != nil {
+		ctx.JSON(http.StatusInternalServerError, err1.Error())
+		return
+	}
+	if !status {
+		ctx.JSON(http.StatusBadRequest, "token expired")
+		return
+	}
+
 	var reply res
 	var insd *models.InstructorDetails
 	err := ctx.BindJSON(&insd)
@@ -39,7 +62,23 @@ func (h *Handler) InstructorInfoHandlers(ctx *gin.Context) {
 }
 
 func (h *Handler) RetrieveInstructorDetails(ctx *gin.Context) {
-	var rid []*models.InstructorDetails
+	ctx.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
+	cookie_token, err1 := ctx.Cookie("token")
+	if err1 != nil {
+		ctx.JSON(http.StatusInternalServerError, err1.Error())
+		return
+	}
+	if cookie_token == "" {
+		ctx.JSON(http.StatusInternalServerError, "authentication error")
+		return
+	}
+
+	validity, _ := h.service.CheckTokenValidity(uuid.MustParse(cookie_token))
+	if !validity {
+		ctx.JSON(http.StatusInternalServerError, "authentication time-out")
+		return
+	}
+
 	rid, err := h.service.GetInstructorDetails()
 	if err != nil {
 		ctx.JSON(http.StatusInternalServerError, err.Error())
@@ -49,6 +88,26 @@ func (h *Handler) RetrieveInstructorDetails(ctx *gin.Context) {
 }
 
 func (h *Handler) DeleteInstructor(ctx *gin.Context) {
+
+	token, err3 := ctx.Cookie("token")
+	if err3 != nil {
+		ctx.JSON(http.StatusInternalServerError, err3.Error())
+		return
+	}
+	token_id, err4 := uuid.Parse(token)
+	if err4 != nil {
+		ctx.JSON(http.StatusInternalServerError, fmt.Errorf("error parsing uuid").Error())
+		return
+	}
+	status, err1 := h.service.CheckTokenValidity(token_id)
+	if err1 != nil {
+		ctx.JSON(http.StatusInternalServerError, err1.Error())
+		return
+	}
+	if !status {
+		ctx.JSON(http.StatusBadRequest, "token expired")
+		return
+	}
 
 	name := ctx.Param("name")
 	err := h.service.DeleteInstructor(name)
